Equation subsystem
